fyp/requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
Django==3.2.9
django-environ==0.7.0
django-froala-editor==4.0.9
environ==1.0
h11==0.13.0
idna==3.3
outcome==1.1.0
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.21
pyOpenSSL==22.0.0
python-magic==0.4.25
python-magic-bin==0.4.14
pytz==2021.1
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
wsproto==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
